pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.4.2",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pickle-mixin>=1.0.2",
    "pillow>=11.2.0",
    "plotly>=6.0.1",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.0",
]
